backDjangoFinalFronEnd/requirements.txt
django==3.2.6
django-allauth==0.45.0
psycopg2==2.9.1
pycco==0.6.0
djangorestframework==3.12.4
django-filter==2.4.0
django-crispy-forms==1.12.0
reportlab==3.6.2
